requirements.txt
acrcloud==14.10.2020
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
anaconda-anon-usage @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_989ovvvj4i/croot/anaconda-anon-usage_1753980472516/work
anaconda-auth @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_fek_jixkaf/croot/anaconda-cloud-auth-split_1747863794091/work
anaconda-cli-base @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_ee_movszx9/croot/anaconda-cli-base_1741369981622/work
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
archspec @ file:///croot/archspec_1709217642129/work
asyncpg==0.31.0
attrs==25.4.0
boltons @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_1cfrddp8h_/croot/boltons_1751383740352/work
brotlicffi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_05qpoetyrv/croot/brotlicffi_1736182869800/work
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/colorama_1728600957751/work
conda @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_a08klfky3c/croot/conda_1754469521713/work/conda-src
conda-anaconda-telemetry @ file:///croot/conda-anaconda-telemetry_1755883788794/work
conda-anaconda-tos @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_3715p1apol/croot/conda-anaconda-tos_1755123333309/work
conda-content-trust @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/conda-content-trust_1728629046705/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1745607008911/work/src
conda-package-handling @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_b5_vfr2fqi/croot/conda-package-handling_1731369042242/work
conda_package_streaming @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_b617jwseui/croot/conda-package-streaming_1751307514977/work
cryptography==46.0.3
distro @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/distro_1728607015893/work
exceptiongroup==1.3.1
fastapi==0.128.0
ffmpeg-python==0.2.0
fire==0.7.1
frozendict @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/frozendict_1728632786107/work
frozenlist==1.8.0
future==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.188.0
google-auth==2.48.0rc0
google-auth-httplib2==0.3.0
google-genai==1.60.0
googleapis-common-protos==1.72.0
greenlet==3.3.1
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
jaraco.classes @ file:///tmp/build/80754af9/jaraco.classes_1620983179379/work
jaraco.context @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/jaraco.context_1731727236461/work
jaraco.functools @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_71brmde743/croot/jaraco.functools_1740408298605/work
Jinja2==3.1.6
jiter==0.13.0
jsonpatch @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/jsonpatch_1728612838217/work
jsonpointer @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_d0dumz8dua/croot/jsonpointer_1753788448378/work
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keyring @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_22mydo866s/croot/keyring_1745484701540/work
libmambapy @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_bd_xm961h4/croot/mamba-split_1734469689868/work/libmambapy
llama-stack-api==0.4.4
llama_stack==0.4.4
markdown-it-py @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/markdown-it-py_1728596183515/work
MarkupSafe==3.0.3
mdurl @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/mdurl_1728595127906/work
menuinst @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_08pr2zh0q9/croot/menuinst_1753464594579/work
more-itertools @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/more-itertools_1728597775985/work
multidict==6.7.1
ollama==0.6.1
openai==2.16.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_95hd3yozva/croot/packaging_1753775365688/work
pillow==12.1.0
pkce @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/pkce_1728608978818/work
platformdirs @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_f8agv_yepz/croot/platformdirs_1744273051859/work
pluggy @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_70ykkrsb12/croot/pluggy_1733169619735/work
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
psycopg2-binary==2.9.11
pyacrcloud==1.0.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycosat @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_67p9tvgx9q/croot/pycosat_1736868714508/work
pycparser==3.0
pydantic==2.12.5
pydantic-settings @ file:///Users/builder/cbouss/buildout/croot/pydantic-settings_1735844448358/work
pydantic_core==2.41.5
pydub==0.25.1
Pygments @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_d3k0kb020b/croot/pygments_1744664134805/work
PyJWT @ file:///Users/builder/cbouss/buildout/croot/pyjwt_1739486501140/work
pyparsing==3.3.2
PySocks @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/pysocks_1728595444110/work
python-dotenv==1.2.1
python-multipart==0.0.21
python-telegram-bot==22.6
PyYAML==6.0.3
readchar @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/readchar_1728610119198/work
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
rich @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_4dgbejjrrq/croot/rich_1732638982215/work
rpds-py==0.30.0
rsa==4.9.1
ruamel.yaml @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_d2rv6_3_xb/croot/ruamel.yaml_1745960311974/work
ruamel.yaml.clib @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_caqnqpjtlv/croot/ruamel.yaml.clib_1745937161530/work
semver @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_222ghyfbla/croot/semver_1736543271469/work
setuptools==78.1.1
shellingham @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/shellingham_1728605541895/work
sniffio==1.3.1
SQLAlchemy==2.0.46
starlette==0.50.0
tenacity==9.1.2
termcolor==3.3.0
tiktoken==0.12.0
tomli @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_09dnnjpgne/croot/tomli_1753774592611/work
tornado==6.5.4
tqdm @ file:///private/var/folders/c_/qfmhj66j0tn016nkx_th4hxm0000gp/T/abs_b8a_tjze9j/croot/tqdm_1738946058169/work
truststore @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_b1m_rjiwrp/croot/truststore_1753907183193/work
typer @ file:///Users/builder/cbouss/perseverance-python-buildout/croot/typer_1728610437695/work
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.40.0
wcwidth==0.5.3
websockets==15.0.1
wheel==0.45.1
yarl==1.22.0
zipp==3.23.0
zstandard @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_65utj9q8ya/croot/zstandard_1731360545821/work
